fix: moved audio thread guard so it noticed by npm
