+++ packages/react-native-audio-api/common/cpp/audioapi/utils/AudioThreadGuard.h
